backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
pydantic==2.5.3
